archive/chiral_brand_processor/requirements.txt
# Core PDF processing libraries
PyMuPDF==1.23.26
PyPDF2==3.0.1
pdfplumber==0.10.3
pdf2image==1.17.0

# Image processing and generation
Pillow==10.2.0
numpy==1.26.4

# OCR capabilities (optional but recommended)
pytesseract==0.3.10

# PDF generation and manipulation
reportlab==4.0.9

# Progress bars and UI
tqdm==4.66.2
colorama==0.4.6

# Text processing
regex==2023.12.25

# Configuration and file handling
pathlib2==2.3.7

# Logging enhancements
loguru==0.7.2

# Image optimization
Pillow-SIMD==10.2.0.post1  # Optional: faster image processing

# System dependencies (install separately if needed):
# - poppler-utils (for pdf2image)
# - tesseract-ocr (for pytesseract)

# Development and testing dependencies (optional)
pytest==8.0.0
pytest-cov==4.1.0
black==24.2.0
flake8==7.0.0

# Performance monitoring (optional)
memory-profiler==0.61.0
psutil==5.9.8